[손희진] 상품 데이터 추가. 상점 작업 중

+++ Assets/Scripts/Data/Ingredient.cs
@@ -6,7 +6,7 @@ public partial class Ingredient
     /// <summary>
     /// ID
     /// </summary>
-    public ushort id;
+    public int id;
     /// <summary>
     /// 인벤토리 데이터 ID
     /// </summary>
